refactor: nits
